refactor: (#205) api 폴더에 있던 wus 폴더 삭제 및 파일 이동
